requirements/base.txt
# base requirements

cmake
Django==2.2.8
django-braces==1.13.0
django-taggit==0.23.0
psycopg2
django-oauth-toolkit==1.2.0
mistune==0.8.4
djangorestframework==3.9.1
redis==3.1.0
Pillow
django-nested-admin==3.1.3
django-cors-headers==2.4.0
whitenoise==4.1.2
django-allauth==0.38.0
faker==1.0.2
face-recognition
tensorflow-cpu
opencv-python-headless
